Updated to support dismissal of OpenAI models
This version removes the need to specify the secret key in the python script and adds it externally. Another solution would have been to use environment variables, but these are not supported by on shinyapps.io
This version also slightly changes the calls to the LLM and the parameters of the sql agent builder as OpenAI has dismissed text-davinci-003. Now using gpt3.5 turbo